backend/requirements.txt
asgiref==3.7.2
Django==4.2
django-cors-headers==4.1.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
PyJWT==2.7.0
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2022.6
